remove unneeded includes and move dtype_dispatcher into a central location
